[REFACTOR] domain 모듈 분리 #82
