+++ umbba-domain/src/main/java/sopt/org/umbba/domain/domain/qna/QuestionType.java
@@ -1,4 +1,4 @@
-package sopt.org.umbba.domain.qna.model;
+package sopt.org.umbba.domain.domain.qna;
 
 import lombok.AccessLevel;
 import lombok.Getter;
