+++ umbba-domain/src/main/java/sopt/org/umbba/domain/domain/user/repository/UserRepository.java
@@ -1,9 +1,9 @@
-package sopt.org.umbba.domain.user.repository;
+package sopt.org.umbba.domain.domain.user.repository;
 
 import org.springframework.data.repository.Repository;
-import sopt.org.umbba.domain.parentchild.model.Parentchild;
-import sopt.org.umbba.domain.user.model.User;
-import sopt.org.umbba.domain.user.social.SocialPlatform;
+import sopt.org.umbba.domain.domain.parentchild.Parentchild;
+import sopt.org.umbba.domain.domain.user.SocialPlatform;
+import sopt.org.umbba.domain.domain.user.User;
 
 import java.util.List;
 import java.util.Optional;
